+++ controllers/watchdog/watchdog.go
@@ -8,6 +8,7 @@ import (
 
 type K8s interface {
 	NewClientset() (kubernetes.Interface, error)
+	GetLabelValue(namespace, label string) (string, error)
 }
 
 type SubController interface {
@@ -18,10 +19,11 @@ type Controller struct {
 	subCtrls []SubController
 }
 
-func NewController(k8s K8s) *Controller {
+func NewController(k8s K8s, nt Notification) *Controller {
 	subCtrls := []SubController{
 		newHPASubController(k8s, log.New("subcontroller", "hpa")),
 		newLimitsSubController(k8s, log.New("subcontroller", "limits")),
+		newServiceSubController(k8s, log.New("subcontroller", "service"), nt),
 	}
 	return &Controller{subCtrls: subCtrls}
 }